src/requirements.txt
astroid==1.6.0
colorama==0.3.9
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.10
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
pip==9.0.1
PyJWT==1.5.3
pylint==1.8.1
pytz==2017.3
setuptools==38.2.4
six==1.11.0
wrapt==1.10.11
